requirements.txt
pillow
pytesseract
pandas
openai
python-dotenv